Add centralized watchlist service
