+++ services/session.py
@@ -2,7 +2,7 @@ import streamlit as st
 
 from config import WATCHLIST_FILE
 from data.portfolio import load_portfolio
-from data.watchlist import load_watchlist, save_watchlist
+from data.watchlist import load_watchlist
 
 
 def init_session_state() -> None:
@@ -35,33 +35,3 @@ def init_session_state() -> None:
 
     if not st.session_state.watchlist and WATCHLIST_FILE.exists():
         st.session_state.watchlist = load_watchlist()
-
-
-def get_watchlist() -> list[str]:
-    """Return the current watchlist from ``st.session_state``."""
-
-    return st.session_state.get("watchlist", [])
-
-
-def add_to_watchlist(ticker: str) -> None:
-    """Add ``ticker`` to the watchlist and persist it."""
-
-    watchlist = st.session_state.setdefault("watchlist", [])
-    symbol = ticker.upper()
-    if symbol and symbol not in watchlist:
-        watchlist.append(symbol)
-        save_watchlist(watchlist)
-
-
-def remove_from_watchlist(symbol: str) -> None:
-    """Remove a ticker symbol from the watchlist in session state and persistent storage."""
-
-    # Load the current list
-    watchlist = get_watchlist()
-    # Remove if present
-    if symbol in watchlist:
-        watchlist.remove(symbol)
-        # Persist the updated list
-        save_watchlist(watchlist)
-        # Update session state if needed
-        st.session_state.watchlist = watchlist
